fix: don't build FUSE paths for Windows
This is a port of https://github.com/googlecloudplatform/cloud-sql-proxy/pull/1400.
